feature_update_database_model
Modified some models/data:
	- CharacterData
	- CountryData
	- ContinentData

Added a prototype from the following controllers:
	- CharacterController
	- CountryController
	- ContinentController

Implemented a configuration class for Hibernate.

Updated application.config.

+++ src/main/java/com/lordofthepushes/facades/user/UserFacade.java
@@ -3,12 +3,11 @@ package com.lordofthepushes.facades.user;
 import com.lordofthepushes.data.UserData;
 import org.springframework.data.domain.Pageable;
 
-
 public interface UserFacade {
     void saveUser(UserData userData);
     void updateUser(UserData userData);
-    void deleteUser(Integer userId);
-    UserData getUserById(Integer userId);
+    void deleteUser(Long userId);
+    UserData getUserById(Long userId);
     Iterable<UserData> getAllUsers();
     Iterable<UserData> getAllUsers(Pageable page);
 }
