+++ src/main/java/com/lordofthepushes/services/user/UserService.java
@@ -6,8 +6,8 @@ import org.springframework.data.domain.Pageable;
 public interface UserService {
     void saveUser(UserData userData);
     void updateUser(UserData userData);
-    void deleteUser(Integer userId);
-    UserData getUserById(Integer userId);
+    void deleteUser(Long userId);
+    UserData getUserById(Long userId);
     Iterable<UserData> getAllUsers();
     Iterable<UserData> getAllUsers(Pageable page);
